Added Authentication to all Endpoints

+++ bellbois/bellbois/functions/deleteBoxItem.js
@@ -1,4 +1,5 @@
 const query = require("../services/SQLService");
+const {authenticateUserContext} = require("../services/authenticationService");
 
 /**
  * A basic Hello World function
@@ -6,5 +7,10 @@ const query = require("../services/SQLService");
  * @returns {object}
  */
 module.exports =  async (ID, context) => {
+  const user = authenticateUserContext(context);
+  if (!user) {
+    throw new Error("Not Authenticated")
+  }
+
   const result = await query("DELETE from `BoxItems` WHERE `BoxItems`.`id` = ?", [ID]);
 };c
